Refactor theme and global state management hooks
- Moved `useTheme` and `useGlobalReducer` hooks to the `shared/hooks` directory for better organization.
- Updated import paths in components and pages to reflect the new hook locations.
- Removed the old implementations of `useTheme` and `useGlobalReducer` from the `hooks` directory.
- Ensured that the `ThemeToggle` component and other relevant components utilize the updated hook paths.
