+++ src/front/pages/Home.jsx
@@ -1,6 +1,6 @@
 import React, { useEffect } from "react"
 import { Link } from "react-router-dom";
-import useGlobalReducer from "../hooks/useGlobalReducer.jsx";
+import useGlobalReducer from "../shared/hooks/useGlobalReducer.jsx";
 import food from "../shared/assets/img/table_food.jpg";
 
 export const Home = () => {
