+++ src/front/shared/hooks/useGlobalReducer.jsx
@@ -1,7 +1,7 @@
 // Import necessary hooks and functions from React.
 import { useContext, useReducer, createContext, useEffect } from "react";
-import storeReducer, { initialStore } from "../store"  // Import the reducer and the initial state.
-import socketService from "../shared/utils/socketService"; // Import socket service for global initialization
+import storeReducer, { initialStore } from "../../store"  // Import the reducer and the initial state.
+import socketService from "../utils/socketService"; // Import socket service for global initialization
 
 // Create a context to hold the global state of the application
 // We will call this global state the "store" to avoid confusion while using local states
